Enhance tile API with SIMD support, blur improvements, and benchmarks

+++ src/simd/include/p10_internal/simd/bitwise_math.hpp
@@ -1,11 +1,12 @@
 #pragma once
 #include <cstddef>
+#include <cstdint>
 
 namespace p10::simd {
 template<size_t b>
-constexpr size_t bitwise_modulo(size_t a) {
+constexpr int64_t bitwise_modulo(int64_t a) {
     static_assert((b & (b - 1)) == 0, "b must be a power of two");
-    return a & (b - 1);
+    return a & static_cast<int64_t>(b - 1);
 }
 
 }  // namespace p10::simd
